+++ src/refactoring/pages/CartPage.tsx
@@ -1,8 +1,9 @@
 import { Coupon, Product } from '../../types.ts';
+import { CartItemManage } from '../components/cart/CartItemManage.tsx';
 import { ProductInfo } from '../components/cart/ProductInfo.tsx';
 import { Layout } from '../components/layout/Layout.tsx';
 import { useCart } from '../hooks/index.ts';
-import { getMaxApplicableDiscount, getMaxDiscount, getRemainingStock } from '../hooks/utils/cartUtils.ts';
+import { getRemainingStock } from '../hooks/utils/cartUtils.ts';
 
 interface Props {
   products: Product[];
@@ -10,7 +11,7 @@ interface Props {
 }
 
 export const CartPage = ({ products, coupons }: Props) => {
-  const { cart, addToCart, removeFromCart, updateQuantity, applyCoupon, calculateTotal, selectedCoupon } = useCart();
+  const { cart, addToCart, updateQuantity, removeFromCart, applyCoupon, calculateTotal, selectedCoupon } = useCart();
 
   const { totalBeforeDiscount, totalAfterDiscount, totalDiscount } = calculateTotal();
 
@@ -33,45 +34,19 @@ export const CartPage = ({ products, coupons }: Props) => {
           })}
         </div>
       </div>
+
       <div>
         <h2 className="text-2xl font-semibold mb-4">장바구니 내역</h2>
 
         <div className="space-y-2">
           {cart.map((item) => {
-            const appliedDiscount = getMaxApplicableDiscount(item);
             return (
-              <div key={item.product.id} className="flex justify-between items-center bg-white p-3 rounded shadow">
-                <div>
-                  <span className="font-semibold">{item.product.name}</span>
-                  <br />
-                  <span className="text-sm text-gray-600">
-                    {item.product.price}원 x {item.quantity}
-                    {appliedDiscount > 0 && (
-                      <span className="text-green-600 ml-1">({(appliedDiscount * 100).toFixed(0)}% 할인 적용)</span>
-                    )}
-                  </span>
-                </div>
-                <div>
-                  <button
-                    onClick={() => updateQuantity(item.product.id, item.quantity - 1)}
-                    className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
-                  >
-                    -
-                  </button>
-                  <button
-                    onClick={() => updateQuantity(item.product.id, item.quantity + 1)}
-                    className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
-                  >
-                    +
-                  </button>
-                  <button
-                    onClick={() => removeFromCart(item.product.id)}
-                    className="bg-red-500 text-white px-2 py-1 rounded hover:bg-red-600"
-                  >
-                    삭제
-                  </button>
-                </div>
-              </div>
+              <CartItemManage
+                key={item.product.id}
+                item={item}
+                updateQuantity={updateQuantity}
+                removeFromCart={removeFromCart}
+              />
             );
           })}
         </div>
